refactor: check root in stat
1. don't run check root when building Backend, run it only when
   Backend::stat instead
2. move messages structs to `message.rs`
3. apply some advices in review
