+++ .github/workflows/service_test_webhdfs.yml
@@ -1,0 +1,50 @@
+name: Service Test WebHDFS
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+    paths-ignore:
+      - "docs/**"
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}-${{ github.event_name }}
+  cancel-in-progress: true
+
+jobs:
+  hdfs:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        hdfs-version: ["2.10.1", "3.2.3", "3.3.2"]
+    steps:
+      - uses: actions/checkout@v3
+
+      - name: Checkout python env
+        uses: actions/setup-python@v4
+        with:
+          python-version: "3.8"
+      - name: Checkout java env
+        uses: actions/setup-java@v3
+        with:
+          distribution: temurin
+          java-version: "11"
+      - name: Setup-hdfs env
+        uses: beyondstorage/setup-hdfs@master
+        with:
+          hdfs-version: ${{ matrix.hdfs-version }}
+
+      - uses: Swatinem/rust-cache@v2
+      - name: Test
+        shell: bash
+        continue-on-error: true
+        run: cargo test hdfs --features compress,services-hdfs -- --show-output
+        env:
+          RUST_BACKTRACE: full
+          RUST_LOG: debug
+          OPENDAL_WEBHDFS_TEST: on
+          OPENDAL_WEBHDFS_ROOT: /
+          OPENDAL_WEBHDFS_ENDPOINT: http://${{ env.WEBHDFS_NAMENODE_ADDR }}
