+++ src/scheme.rs
@@ -73,8 +73,8 @@ pub enum Scheme {
     S3,
     /// [webdav][crate::services::Webdav]: WebDAV support.
     Webdav,
-    /// [webhdfs][crate::services::webhdfs]: WebHDFS RESTful API Services
-    WebHdfs,
+    /// [webhdfs][crate::services::WebHdfs]: WebHDFS RESTful API Services
+    Webhdfs,
     /// Custom that allow users to implement services outside of OpenDAL.
     ///
     /// # NOTE
@@ -135,7 +135,7 @@ impl FromStr for Scheme {
             "s3" => Ok(Scheme::S3),
             "oss" => Ok(Scheme::Oss),
             "webdav" => Ok(Scheme::Webdav),
-            "webhdfs" => Ok(Scheme::WebHdfs),
+            "webhdfs" => Ok(Scheme::Webhdfs),
             _ => Ok(Scheme::Custom(Box::leak(s.into_boxed_str()))),
         }
     }
@@ -170,7 +170,7 @@ impl From<Scheme> for &'static str {
             Scheme::S3 => "s3",
             Scheme::Oss => "oss",
             Scheme::Webdav => "webdav",
-            Scheme::WebHdfs => "webhdfs",
+            Scheme::Webhdfs => "webhdfs",
             Scheme::Custom(v) => v,
         }
     }
