+++ src/services/webhdfs/backend.rs
@@ -24,56 +24,18 @@ use http::Response;
 use http::StatusCode;
 use log::debug;
 use log::warn;
-use serde::Deserialize;
+use tokio::sync::OnceCell;
 
-use super::auth::WebHdfsAuth;
 use super::dir_stream::DirStream;
-use super::error::blocking_parse_error;
-use super::error::blocking_resp_to_str;
 use super::error::parse_error;
-use super::percent_encode_path;
-use crate::raw::build_abs_path;
-use crate::raw::input;
-use crate::raw::new_request_build_error;
-use crate::raw::normalize_root;
-use crate::raw::parse_into_object_metadata;
-use crate::raw::Accessor;
-use crate::raw::AccessorCapability;
-use crate::raw::AccessorMetadata;
-use crate::raw::AsyncBody;
-use crate::raw::Body;
-use crate::raw::BytesRange;
-use crate::raw::HttpClient;
-use crate::raw::IncomingAsyncBody;
-use crate::raw::ObjectPage;
-use crate::raw::ObjectPager;
-use crate::raw::RpCreate;
-use crate::raw::RpDelete;
-use crate::raw::RpList;
-use crate::raw::RpRead;
-use crate::raw::RpStat;
-use crate::raw::RpWrite;
-use crate::Builder;
-use crate::Error;
-use crate::ErrorKind;
-use crate::ObjectMetadata;
-use crate::ObjectMode;
-use crate::OpCreate;
-use crate::OpDelete;
-use crate::OpList;
-use crate::OpRead;
-use crate::OpStat;
-use crate::OpWrite;
-use crate::Result;
-use crate::Scheme;
+use super::message::BooleanResp;
+use super::message::FileStatusWrapper;
+use super::message::FileStatusesWrapper;
+use super::message::Redirection;
+use crate::raw::*;
+use crate::*;
 
 const WEBHDFS_DEFAULT_ENDPOINT: &str = "http://127.0.0.1:9870";
-
-type AsyncReq = Request<AsyncBody>;
-type AsyncResp = Response<IncomingAsyncBody>;
-
-type BlockReq = Request<Body>;
-type BlockResp = Response<Body>;
 
 /// WebHDFS's RESTFul API support
 ///
@@ -92,19 +54,17 @@ type BlockResp = Response<Body>;
 /// - `OPENDAL_WEBHDFS_ENDPOINT`
 /// - `OPENDAL_WEBHDFS_USERNAME`
 /// - `OPENDAL_WEBHDFS_DELEGATION`
-/// - `OPENDAL_WEBHDFS_CSRF`
 ///
 /// # Examples
 /// ## Via Environment
 ///
-/// Set the environment variables and then use [`WebHdfs::try_from_env`]:
+/// Set the environment variables and then use [`Webhdfs::try_from_env`]:
 ///
 /// ```shell
 /// export OPENDAL_WEBHDFS_ROOT=/path/to/dir
 /// export OPENDAL_WEBHDFS_DELEGATION=<delegation_token>
 /// export OPENDAL_WEBHDFS_USERNAME=<username>
 /// export OPENDAL_WEBHDFS_ENDPOINT=localhost:50070
-/// export OPENDAL_WEBHDFS_CSRF=X-XSRF-HEADER
 /// ```
 /// ```no_run
 /// use std::sync::Arc;
@@ -115,7 +75,7 @@ type BlockResp = Response<Body>;
 ///
 /// #[tokio::main]
 /// async fn main() -> Result<()> {
-///     let op: Operator = Operator::from_env(Scheme::WebHdfs)?;
+///     let op: Operator = Operator::from_env(Scheme::Webhdfs)?;
 ///     let _: Object = op.object("test_file");
 ///     Ok(())
 /// }
@@ -125,13 +85,13 @@ type BlockResp = Response<Body>;
 /// use std::sync::Arc;
 ///
 /// use anyhow::Result;
-/// use opendal::services::WebHdfs;
+/// use opendal::services::Webhdfs;
 /// use opendal::Object;
 /// use opendal::Operator;
 ///
 /// #[tokio::main]
 /// async fn main() -> Result<()> {
-///     let mut builder = WebHdfs::default();
+///     let mut builder = Webhdfs::default();
 ///     // set the root for s3, all operations will happend under this root
 ///     //
 ///     // Note:
@@ -143,7 +103,6 @@ type BlockResp = Response<Body>;
 ///     // set the endpoint of webhdfs namenode, controled by dfs.namenode.http-address
 ///     // default is http://127.0.0.1:9870
 ///     builder.endpoint("http://127.0.0.1:9870");
-
 ///     // set the delegation_token for builder
 ///     builder.delegation("delegation_token");
 ///     // or set the username for builder
@@ -152,10 +111,6 @@ type BlockResp = Response<Body>;
 ///     builder.doas("proxy_user");
 ///     // if no delegation token, username and proxy user are set
 ///     // the backend will query without authentications
-///
-///     // custom CSRF header is also customizable
-///     // default is X-XSRF-HEADER
-///     builder.csrf("X-XSRF-HEADER");
 ///
 ///     let op: Operator = Operator::new(builder.build()?);
 ///
@@ -166,18 +121,16 @@ type BlockResp = Response<Body>;
 /// }
 /// ```
 #[derive(Default, Clone)]
-pub struct WebHdfsBuilder {
+pub struct WebhdfsBuilder {
     root: Option<String>,
     endpoint: Option<String>,
     delegation: Option<String>,
     user: Option<String>,
     // proxy user
     doas: Option<String>,
-    // custom CSRF prevention header name
-    csrf: Option<String>,
 }
 
-impl Debug for WebHdfsBuilder {
+impl Debug for WebhdfsBuilder {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let mut ds = f.debug_struct("Builder");
         ds.field("root", &self.root)
@@ -191,14 +144,11 @@ impl Debug for WebHdfsBuilder {
         if self.doas.is_some() {
             ds.field("doas", &"<redacted>");
         }
-        if self.csrf.is_some() {
-            ds.field("csrf", &"<redacted>");
-        }
         ds.finish()
     }
 }
 
-impl WebHdfsBuilder {
+impl WebhdfsBuilder {
     /// Set the working directory of this backend
     ///
     /// All operations will happen under this root
@@ -269,26 +219,30 @@ impl WebHdfsBuilder {
         }
         self
     }
+}
 
-    /// Set th CSRF prevention header name
-    /// refer to [CSRF prevention](https://hadoop.apache.org/docs/r3.3.3/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Cross-Site_Request_Forgery_Prevention) for detail
-    ///
-    /// # Note:
-    /// If not set, the default value is `X-XSRF-HEADER`
-    pub fn csrf(&mut self, csrf: &str) -> &mut Self {
-        if !csrf.is_empty() {
-            self.csrf = Some(csrf.to_string());
+impl WebhdfsBuilder {
+    fn auth_str(&mut self) -> Option<String> {
+        if let Some(dt) = self.delegation.take() {
+            return Some(format!("delegation_token={dt}"));
         }
-        self
+        if let Some(user) = self.user.take() {
+            let token = match self.doas.take() {
+                Some(doas) => format!("user.name={user}&doas={doas}"),
+                None => format!("user.name={user}"),
+            };
+            return Some(token);
+        }
+        None
     }
 }
 
-impl Builder for WebHdfsBuilder {
-    const SCHEME: Scheme = Scheme::WebHdfs;
-    type Accessor = WebHdfsBackend;
+impl Builder for WebhdfsBuilder {
+    const SCHEME: Scheme = Scheme::Webhdfs;
+    type Accessor = WebhdfsBackend;
 
     fn from_map(map: HashMap<String, String>) -> Self {
-        let mut builder = WebHdfsBuilder::default();
+        let mut builder = WebhdfsBuilder::default();
 
         for (k, v) in map.iter() {
             let v = v.as_str();
@@ -298,7 +252,6 @@ impl Builder for WebHdfsBuilder {
                 "delegation" => builder.delegation(v),
                 "user" => builder.user(v),
                 "doas" => builder.doas(v),
-                "csrf" => builder.csrf(v),
                 _ => continue,
             };
         }
@@ -325,40 +278,18 @@ impl Builder for WebHdfsBuilder {
         };
         debug!("backend use endpoint {}", endpoint);
 
-        let csrf = match self.csrf.take() {
-            Some(c) => c,
-            None => "X-XSRF-HEADER".to_string(),
-        };
-
-        let auth = {
-            if let Some(token) = self.delegation.take() {
-                Some(WebHdfsAuth::new_delegation(&token))
-            } else if let Some(username) = self.user.take() {
-                let doas = if let Some(doas) = self.doas.take() {
-                    doas
-                } else {
-                    String::new()
-                };
-                Some(WebHdfsAuth::new_user(&username, &doas))
-            } else {
-                debug!("neither delegation token nor username is set");
-                None
-            }
-        };
-        let auth_str = auth.map(|s| s.auth_str());
-
+        let auth = self.auth_str();
         let client = HttpClient::new()?;
 
-        let backend = WebHdfsBackend {
+        let backend = WebhdfsBackend {
             root: root.clone(),
             endpoint,
-            auth: auth_str,
+            auth,
             client,
-            csrf,
+            root_checker: OnceCell::new(),
         };
 
         debug!("checking working directory: {}", root);
-        backend.check_root()?;
 
         Ok(backend)
     }
@@ -366,22 +297,22 @@ impl Builder for WebHdfsBuilder {
 
 /// Backend for WebHDFS service
 #[derive(Debug, Clone)]
-pub struct WebHdfsBackend {
+pub struct WebhdfsBackend {
     root: String,
     endpoint: String,
     client: HttpClient,
     auth: Option<String>,
-    csrf: String,
+    root_checker: OnceCell<()>,
 }
 
-impl WebHdfsBackend {
+impl WebhdfsBackend {
     async fn create_or_write_object_req(
         &self,
         path: &str,
         size: Option<u64>,
         content_type: Option<&str>,
         body: AsyncBody,
-    ) -> Result<AsyncReq> {
+    ) -> Result<Request<AsyncBody>> {
         let p = build_abs_path(&self.root, path);
         let op = if path.ends_with('/') {
             "MKDIRS"
@@ -399,7 +330,6 @@ impl WebHdfsBackend {
         }
 
         let req = Request::put(&url)
-            .header(self.csrf.as_str(), "")
             .body(AsyncBody::Empty)
             .map_err(new_request_build_error)?;
         // mkdir does not redirect
@@ -413,7 +343,7 @@ impl WebHdfsBackend {
             .await
     }
 
-    async fn open_object_req(&self, path: &str, range: &BytesRange) -> Result<AsyncReq> {
+    async fn open_object_req(&self, path: &str, range: &BytesRange) -> Result<Request<AsyncBody>> {
         let p = build_abs_path(&self.root, path);
         let mut url = format!(
             "{}/webhdfs/v1/{}?op=OPEN&noredirect=true",
@@ -424,9 +354,9 @@ impl WebHdfsBackend {
             url = url + "&" + auth;
         }
 
-        // make a WebHdfs compatible bytes range
+        // make a Webhdfs compatible bytes range
         //
-        // WebHdfs does not support read from end
+        // Webhdfs does not support read from end
         // have to solve manually
         let range = match (range.offset(), range.size()) {
             // avoiding reading the whole file
@@ -458,50 +388,13 @@ impl WebHdfsBackend {
         }
 
         let req = Request::get(&url)
-            .header(self.csrf.as_str(), "")
             .body(AsyncBody::Empty)
             .map_err(new_request_build_error)?;
 
         Ok(req)
     }
 
-    fn status_object_req<B>(&self, path: &str, body: B) -> Result<Request<B>> {
-        let p = build_abs_path(&self.root, path);
-        let mut url = format!(
-            "{}/webhdfs/v1/{}?op=GETFILESTATUS",
-            self.endpoint,
-            percent_encode_path(&p),
-        );
-        if let Some(auth) = &self.auth {
-            url = url + "&" + auth;
-        }
-
-        let req = Request::get(&url);
-        req.header(self.csrf.as_str(), "")
-            .body(body)
-            .map_err(new_request_build_error)
-    }
-
-    fn delete_object_req(&self, path: &str) -> Result<AsyncReq> {
-        let p = build_abs_path(&self.root, path);
-        let mut url = format!(
-            "{}/webhdfs/v1/{}?op=DELETE&recursive=false",
-            self.endpoint,
-            percent_encode_path(&p),
-        );
-        if let Some(auth) = &self.auth {
-            url = url + "&" + auth;
-        }
-
-        let req = Request::delete(&url);
-        let req = req
-            .header(self.csrf.as_str(), "")
-            .body(AsyncBody::Empty)
-            .map_err(new_request_build_error)?;
-        Ok(req)
-    }
-
-    fn list_object_req(&self, path: &str) -> Result<AsyncReq> {
+    fn list_object_req(&self, path: &str) -> Result<Request<AsyncBody>> {
         let p = build_abs_path(&self.root, path);
         let mut url = format!(
             "{}/webhdfs/v1/{}?op=LISTSTATUS",
@@ -514,43 +407,22 @@ impl WebHdfsBackend {
 
         let req = Request::get(&url);
         let req = req
-            .header(self.csrf.as_str(), "")
             .body(AsyncBody::Empty)
             .map_err(new_request_build_error)?;
         Ok(req)
     }
 }
 
-impl WebHdfsBackend {
-    // blocking create or write object request
-    fn blocking_create_root_req(&self) -> Result<BlockReq> {
-        let p = &self.root;
-        let op = "MKDIRS";
-        let mut url = format!(
-            "{}/webhdfs/v1/{}?op={}&overwrite=true&noredirect=true",
-            self.endpoint,
-            percent_encode_path(p),
-            op,
-        );
-        if let Some(auth) = &self.auth {
-            url = url + "&" + auth;
-        }
-
-        let req = Request::put(&url)
-            .header(self.csrf.as_str(), "")
-            .body(Body::Empty)
-            .map_err(new_request_build_error)?;
-        // mkdir does not redirect
-        Ok(req)
-    }
-}
-
-impl WebHdfsBackend {
+impl WebhdfsBackend {
     /// get object from webhdfs
     /// # Note
     /// looks like webhdfs doesn't support range request from file end.
     /// so if we want to read the tail of object, the whole object should be transfered.
-    async fn get_object(&self, path: &str, range: BytesRange) -> Result<AsyncResp> {
+    async fn get_object(
+        &self,
+        path: &str,
+        range: BytesRange,
+    ) -> Result<Response<IncomingAsyncBody>> {
         let req = self.open_object_req(path, &range).await?;
         let resp = self.client.send_async(req).await?;
 
@@ -565,27 +437,55 @@ impl WebHdfsBackend {
         self.client.send_async(redirected).await
     }
 
-    async fn status_object(&self, path: &str) -> Result<AsyncResp> {
-        let req = self.status_object_req(path, AsyncBody::Empty)?;
+    async fn status_object(&self, path: &str) -> Result<Response<IncomingAsyncBody>> {
+        let p = build_abs_path(&self.root, path);
+        let mut url = format!(
+            "{}/webhdfs/v1/{}?op=GETFILESTATUS",
+            self.endpoint,
+            percent_encode_path(&p),
+        );
+        if let Some(auth) = &self.auth {
+            url = url + "&" + auth;
+        }
+
+        let req = Request::get(&url);
+        let req = req
+            .body(AsyncBody::Empty)
+            .map_err(new_request_build_error)?;
+
         self.client.send_async(req).await
     }
 
-    async fn delete_object(&self, path: &str) -> Result<AsyncResp> {
-        let req = self.delete_object_req(path)?;
+    async fn delete_object(&self, path: &str) -> Result<Response<IncomingAsyncBody>> {
+        let p = build_abs_path(&self.root, path);
+        let mut url = format!(
+            "{}/webhdfs/v1/{}?op=DELETE&recursive=false",
+            self.endpoint,
+            percent_encode_path(&p),
+        );
+        if let Some(auth) = &self.auth {
+            url = url + "&" + auth;
+        }
+
+        let req = Request::delete(&url);
+        let req = req
+            .body(AsyncBody::Empty)
+            .map_err(new_request_build_error)?;
+
         self.client.send_async(req).await
     }
 
     /// get redirect destination from 307 TEMPORARY_REDIRECT http response
-    async fn follow_redirect(&self, resp: AsyncResp) -> Result<String> {
+    async fn follow_redirect(&self, resp: Response<IncomingAsyncBody>) -> Result<String> {
         let bs = resp.into_body().bytes().await.map_err(|e| {
             Error::new(ErrorKind::Unexpected, "redirection receive fail")
-                .with_context("service", Scheme::WebHdfs)
+                .with_context("service", Scheme::Webhdfs)
                 .set_source(e)
         })?;
         let loc = serde_json::from_reader::<_, Redirection>(bs.reader())
             .map_err(|e| {
                 Error::new(ErrorKind::Unexpected, "redirection fail")
-                    .with_context("service", Scheme::WebHdfs)
+                    .with_context("service", Scheme::Webhdfs)
                     .set_permanent()
                     .set_source(e)
             })?
@@ -595,30 +495,25 @@ impl WebHdfsBackend {
     }
 }
 
-impl WebHdfsBackend {
-    fn blocking_status_object(&self, path: &str) -> Result<BlockResp> {
-        let req = self.status_object_req(path, Body::Empty)?;
-        self.client.send(req)
-    }
-}
-
-impl WebHdfsBackend {
-    async fn get_along_redirect(&self, redirection: AsyncResp) -> Result<AsyncReq> {
+impl WebhdfsBackend {
+    async fn get_along_redirect(
+        &self,
+        redirection: Response<IncomingAsyncBody>,
+    ) -> Result<Request<AsyncBody>> {
         let redirect = self.follow_redirect(redirection).await?;
 
         Request::get(redirect)
-            .header(self.csrf.as_str(), "")
             .body(AsyncBody::Empty)
             .map_err(new_request_build_error)
     }
 
     async fn put_along_redirect(
         &self,
-        resp: AsyncResp,
+        resp: Response<IncomingAsyncBody>,
         size: Option<u64>,
         content_type: Option<&str>,
         body: AsyncBody,
-    ) -> Result<AsyncReq> {
+    ) -> Result<Request<AsyncBody>> {
         let redirect = self.follow_redirect(resp).await?;
 
         let mut req = Request::put(redirect);
@@ -628,16 +523,14 @@ impl WebHdfsBackend {
         if let Some(content_type) = content_type {
             req = req.header(CONTENT_TYPE, content_type);
         }
-        req.header(self.csrf.as_str(), "")
-            .body(body)
-            .map_err(new_request_build_error)
+        req.body(body).map_err(new_request_build_error)
     }
 
     fn consume_success_mkdir(&self, path: &str, parts: Parts, body: &str) -> Result<RpCreate> {
         let mkdir_rsp = serde_json::from_str::<BooleanResp>(body).map_err(|e| {
             Error::new(ErrorKind::Unexpected, "cannot parse mkdir response")
                 .set_temporary()
-                .with_context("service", Scheme::WebHdfs)
+                .with_context("service", Scheme::Webhdfs)
                 .with_context("response", format!("{parts:?}"))
                 .set_source(e)
         })?;
@@ -653,64 +546,23 @@ impl WebHdfsBackend {
     }
 }
 
-impl WebHdfsBackend {
-    fn stat_root(&self) -> Result<ObjectMode> {
-        let resp = self.blocking_status_object("/")?;
-        let status = resp.status();
-        match status {
-            StatusCode::OK => {
-                let (_, s) = blocking_resp_to_str(resp)?;
-
-                let file_status = serde_json::from_str::<FileStatusWrapper>(&s)
-                    .map_err(|e| {
-                        Error::new(ErrorKind::Unexpected, "cannot parse returned json")
-                            .with_context("service", Scheme::WebHdfs)
-                            .set_source(e)
-                    })?
-                    .file_status;
-                let status_meta: ObjectMetadata = file_status.try_into()?;
-
-                Ok(status_meta.mode())
-            }
-            _ => Err(blocking_parse_error(resp)?),
-        }
-    }
-
-    fn create_root(&self) -> Result<RpCreate> {
-        let path = "/";
-        let req = self.blocking_create_root_req()?;
-        let resp = self.client.send(req)?;
-
-        let status = resp.status();
-
-        // WebHDFS's has a two-step create/append to prevent clients to send out
-        // data before creating it.
-        // According to the redirect policy of `reqwest` HTTP Client we are using,
-        // the redirection should be done automatically.
-        match status {
-            StatusCode::CREATED | StatusCode::OK => {
-                let (parts, body) = blocking_resp_to_str(resp)?;
-                self.consume_success_mkdir(path, parts, &body)
-            }
-            _ => Err(blocking_parse_error(resp)?),
-        }
-    }
-
-    fn check_root(&self) -> Result<()> {
-        match self.stat_root() {
-            Ok(mode) => {
+impl WebhdfsBackend {
+    async fn check_root(&self) -> Result<()> {
+        match self.stat("/", OpStat::default()).await {
+            Ok(rp) => {
+                let mode = rp.into_metadata().mode();
                 if !mode.is_dir() {
                     warn!("working directory is occupied!");
                     return Err(
                         Error::new(ErrorKind::BackendConfigInvalid, "root is occupied!")
-                            .with_context("service", Scheme::WebHdfs),
+                            .with_context("service", Scheme::Webhdfs),
                     );
                 }
             }
             Err(e) => match e.kind() {
                 ErrorKind::ObjectNotFound => {
                     debug!("working directory does not exists, creating...");
-                    self.create_root()?;
+                    self.create("/", OpCreate::new(ObjectMode::DIR)).await?;
                 }
                 _ => return Err(e),
             },
@@ -721,13 +573,13 @@ impl WebHdfsBackend {
 }
 
 #[async_trait]
-impl Accessor for WebHdfsBackend {
+impl Accessor for WebhdfsBackend {
     type Reader = IncomingAsyncBody;
     type BlockingReader = ();
 
     fn metadata(&self) -> AccessorMetadata {
         let mut am = AccessorMetadata::default();
-        am.set_scheme(Scheme::WebHdfs)
+        am.set_scheme(Scheme::Webhdfs)
             .set_root(&self.root)
             .set_capabilities(
                 AccessorCapability::Read | AccessorCapability::Write | AccessorCapability::List,
@@ -784,7 +636,7 @@ impl Accessor for WebHdfsBackend {
                 Ok((RpRead::with_metadata(meta), resp.into_body()))
             }
             StatusCode::NOT_FOUND => Err(Error::new(ErrorKind::ObjectNotFound, "object not found")
-                .with_context("service", Scheme::WebHdfs)),
+                .with_context("service", Scheme::Webhdfs)),
             _ => Err(parse_error(resp).await?),
         }
     }
@@ -811,6 +663,11 @@ impl Accessor for WebHdfsBackend {
     }
 
     async fn stat(&self, path: &str, _: OpStat) -> Result<RpStat> {
+        // if root exists and is a directory, stat will be ok
+        self.root_checker
+            .get_or_try_init(|| async { self.check_root().await })
+            .await?;
+
         let resp = self.status_object(path).await?;
         let status = resp.status();
         match status {
@@ -821,7 +678,7 @@ impl Accessor for WebHdfsBackend {
                 let file_status = serde_json::from_reader::<_, FileStatusWrapper>(body_bs.reader())
                     .map_err(|e| {
                         Error::new(ErrorKind::Unexpected, "cannot parse returned json")
-                            .with_context("service", Scheme::WebHdfs)
+                            .with_context("service", Scheme::Webhdfs)
                             .set_source(e)
                     })?
                     .file_status;
@@ -861,7 +718,7 @@ impl Accessor for WebHdfsBackend {
                     serde_json::from_reader::<_, FileStatusesWrapper>(body_bs.reader())
                         .map_err(|e| {
                             Error::new(ErrorKind::Unexpected, "cannot parse returned json")
-                                .with_context("service", Scheme::WebHdfs)
+                                .with_context("service", Scheme::Webhdfs)
                                 .set_source(e)
                         })?
                         .file_statuses
@@ -876,181 +733,5 @@ impl Accessor for WebHdfsBackend {
             }
             _ => Err(parse_error(resp).await?),
         }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct Redirection {
-    pub location: String,
-}
-
-#[derive(Debug, Deserialize)]
-struct BooleanResp {
-    pub boolean: bool,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct FileStatusWrapper {
-    pub file_status: FileStatus,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-struct FileStatusesWrapper {
-    pub file_statuses: FileStatuses,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "PascalCase")]
-pub(super) struct FileStatuses {
-    pub file_status: Vec<FileStatus>,
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub(super) struct FileStatus {
-    pub length: u64,
-    pub modification_time: i64,
-
-    pub path_suffix: String,
-    #[serde(rename = "type")]
-    pub ty: FileStatusType,
-}
-
-impl TryFrom<FileStatus> for ObjectMetadata {
-    type Error = Error;
-    fn try_from(value: FileStatus) -> Result<Self> {
-        let mut meta = match value.ty {
-            FileStatusType::Directory => ObjectMetadata::new(ObjectMode::DIR),
-            FileStatusType::File => ObjectMetadata::new(ObjectMode::FILE),
-        };
-        let till_now = time::Duration::milliseconds(value.modification_time);
-        let last_modified = time::OffsetDateTime::UNIX_EPOCH
-            .checked_add(till_now)
-            .ok_or(Error::new(
-                ErrorKind::Unexpected,
-                "last modification overflowed!",
-            ))?;
-        meta.set_last_modified(last_modified)
-            .set_content_length(value.length);
-        Ok(meta)
-    }
-}
-
-#[derive(Debug, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "UPPERCASE")]
-pub(super) enum FileStatusType {
-    Directory,
-    File,
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_file_statuses() {
-        let json = r#"{"Location":"http://<DATANODE>:<PORT>/webhdfs/v1/<PATH>?op=CREATE..."}"#;
-        let redir: Redirection = serde_json::from_str(json).expect("must success");
-        assert_eq!(
-            redir.location,
-            "http://<DATANODE>:<PORT>/webhdfs/v1/<PATH>?op=CREATE..."
-        );
-    }
-
-    #[test]
-    fn test_file_status() {
-        let json = r#"
-{
-  "FileStatus":
-  {
-    "accessTime"      : 0,
-    "blockSize"       : 0,
-    "group"           : "supergroup",
-    "length"          : 0,
-    "modificationTime": 1320173277227,
-    "owner"           : "webuser",
-    "pathSuffix"      : "",
-    "permission"      : "777",
-    "replication"     : 0,
-    "type"            : "DIRECTORY"
-  }
-}
-"#;
-        let status: FileStatusWrapper = serde_json::from_str(json).expect("must success");
-        assert_eq!(status.file_status.length, 0);
-        assert_eq!(status.file_status.modification_time, 1320173277227);
-        assert_eq!(status.file_status.path_suffix, "");
-        assert_eq!(status.file_status.ty, FileStatusType::Directory);
-    }
-
-    #[tokio::test]
-    async fn test_list_empty() {
-        let json = r#"
-    {
-        "FileStatuses": {"FileStatus":[]}
-    }
-        "#;
-        let file_statuses = serde_json::from_str::<FileStatusesWrapper>(json)
-            .expect("must success")
-            .file_statuses
-            .file_status;
-        assert!(file_statuses.is_empty());
-    }
-
-    #[tokio::test]
-    async fn test_list_status() {
-        let json = r#"
-{
-  "FileStatuses":
-  {
-    "FileStatus":
-    [
-      {
-        "accessTime"      : 1320171722771,
-        "blockSize"       : 33554432,
-        "group"           : "supergroup",
-        "length"          : 24930,
-        "modificationTime": 1320171722771,
-        "owner"           : "webuser",
-        "pathSuffix"      : "a.patch",
-        "permission"      : "644",
-        "replication"     : 1,
-        "type"            : "FILE"
-      },
-      {
-        "accessTime"      : 0,
-        "blockSize"       : 0,
-        "group"           : "supergroup",
-        "length"          : 0,
-        "modificationTime": 1320895981256,
-        "owner"           : "szetszwo",
-        "pathSuffix"      : "bar",
-        "permission"      : "711",
-        "replication"     : 0,
-        "type"            : "DIRECTORY"
-      }
-    ]
-  }
-}
-            "#;
-
-        let file_statuses = serde_json::from_str::<FileStatusesWrapper>(json)
-            .expect("must success")
-            .file_statuses
-            .file_status;
-
-        let mut pager = DirStream::new("listing/directory", file_statuses);
-        let mut entries = vec![];
-        while let Some(oes) = pager.next_page().await.expect("must success") {
-            entries.extend(oes);
-        }
-        assert_eq!(entries.len(), 2);
-        assert_eq!(entries[0].path(), "listing/directory/a.patch");
-        assert_eq!(entries[0].mode(), ObjectMode::FILE);
-        assert_eq!(entries[1].path(), "listing/directory/bar/");
-        assert_eq!(entries[1].mode(), ObjectMode::DIR);
     }
 }
