+++ src/services/webhdfs/error.rs
@@ -12,19 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::io::Read;
-
 use http::response::Parts;
 use http::Response;
 use http::StatusCode;
 use serde::Deserialize;
 
-use crate::raw::Body;
-use crate::raw::IncomingAsyncBody;
-use crate::Error;
-use crate::ErrorKind;
-use crate::Result;
-use crate::Scheme;
+use crate::raw::*;
+use crate::*;
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "PascalCase")]
@@ -49,36 +43,14 @@ pub(super) async fn parse_error(resp: Response<IncomingAsyncBody>) -> Result<Err
     parse_error_msg(parts, &s)
 }
 
-pub(super) fn blocking_parse_error(resp: Response<Body>) -> Result<Error> {
-    let (p, s) = blocking_resp_to_str(resp)?;
-    parse_error_msg(p, &s)
-}
-
-pub(super) fn blocking_resp_to_str(resp: Response<Body>) -> Result<(Parts, String)> {
-    let (parts, mut body) = resp.into_parts();
-
-    let mut s = String::new();
-    body.read_to_string(&mut s).map_err(|e| {
-        Error::new(
-            ErrorKind::Unexpected,
-            &format!("got status {} and unparsable body", parts.status.as_str()),
-        )
-        .with_context("service", Scheme::WebHdfs)
-        .with_context("response", format!("{parts:?}"))
-        .set_temporary()
-        .set_source(e)
-    })?;
-    Ok((parts, s))
-}
-
 fn parse_error_msg(parts: Parts, body: &str) -> Result<Error> {
     let (kind, retryable) = match parts.status {
         StatusCode::NOT_FOUND => (ErrorKind::ObjectNotFound, false),
         StatusCode::UNAUTHORIZED | StatusCode::FORBIDDEN => {
             (ErrorKind::ObjectPermissionDenied, false)
         }
-        StatusCode::BAD_REQUEST => (ErrorKind::Unsupported, false),
-        StatusCode::INTERNAL_SERVER_ERROR
+        StatusCode::BAD_REQUEST
+        | StatusCode::INTERNAL_SERVER_ERROR
         | StatusCode::BAD_GATEWAY
         | StatusCode::SERVICE_UNAVAILABLE
         | StatusCode::GATEWAY_TIMEOUT => (ErrorKind::Unexpected, true),
